Update FINAL_CODE
In this part of our code we demonstrated the low button state and set up the default situation and format. We also ran into some issues here as our sensor was not working properly {as mentioned in the video}. We talked with Prof. Anima in regards to this and she mentioned trying to use a different call to the library function, however we then ran into some other issues. (Jake)